+++ packages/eslint/src/lib/consts.ts
@@ -1,0 +1,10 @@
+export const userIgnorePatterns = [
+  '!**/*',
+  'dist',
+  'node_modules',
+  '.idea',
+  '.yarn',
+  'build',
+  'tmp',
+  'coverage'
+]