+++ packages/eslint/src/lib/types.ts
@@ -1,6 +1,4 @@
-import { configure, presets } from 'eslint-kit'
-
-import { EslintKitPresetsOptions } from './presets'
+import { configure , presets } from 'eslint-kit'
 
 export type Options = Parameters<typeof configure>[0]
 
@@ -8,8 +6,12 @@ export type KitPreset = Options['presets'][number]
 
 export type ValidPresets = keyof typeof presets
 
-export interface CreateEslintConfigOptions extends EslintKitPresetsOptions {
+export interface ConfigOptions {
+  tsconfig?: string
+  extensions?: ValidPresets[]
+  enableImports?: boolean
   root: string
   ignore?: string[]
   allowDebug?: boolean
+  monorepo?: boolean
 }